Refactoring
Quick fix
